refactor(server): Move ip valid func to utils mod
- Helps in keeping `server.py` module light-weight and less cluttered
