+++ src/rupantar/sohoj/utils.py
@@ -1,5 +1,6 @@
 from time import perf_counter
 from logging import getLogger
+from ipaddress import ip_address
 
 logger = getLogger()
 
@@ -39,3 +40,27 @@ def get_func_exec_time(function):
         return og_func_return
 
     return wrap
+
+
+def validate_network_address(interface_address: str) -> bool:
+    """Validate a given network address to check if it is a valid IP address.
+
+    Link-local and multicast addresses are considered to be invalid.
+
+    Note:
+        The link-local addresses = reserved range of 169.254.1.0 to 169.254.254.255, as per the IETF.
+
+    Args:
+      interface_address: Network address to validate
+
+    Returns:
+        bool: True if the given network address is a valid, non-link-local, non-multicast IP address. False otherwise.
+
+    """
+    try:
+        ip = ip_address(interface_address)
+        if ip.is_link_local or ip.is_multicast:
+            return False
+        return True
+    except ValueError:
+        return False
